+++ javaScript/projects/binary-searh-trees/src/components/test.js
@@ -3,12 +3,11 @@ import Icon from '../assets/icons/sharp_home.svg';
 export default {
   print() {
     console.log('print() running from test.js');
-    console.log('testing...');
   },
   render() {
     const div = document.createElement('div');
     const paragraph = document.createElement('p');
-    const foo = "test";
+    const foo = 'test';
 
     const icon = new Image();
     icon.src = Icon;